Move MutationScheduleAutoMode to a new module
