+++ src/nextline_schedule/schema/auto/__init__.py
@@ -1,0 +1,5 @@
+__all__ = [
+    'MutationScheduleAutoMode',
+]
+
+from .mutation import MutationScheduleAutoMode
