+++ src/nextline_schedule/schema/mutation.py
@@ -4,30 +4,10 @@ import strawberry
 from nextline import Nextline
 from strawberry.types import Info
 
-from nextline_schedule.auto import AutoMode
 from nextline_schedule.scheduler import RequestStatement
 
+from .auto import MutationScheduleAutoMode
 from .queue import MutationScheduleQueue
-
-
-async def mutate_turn_on(info: Info) -> bool:
-    auto_mode = info.context['schedule']['auto_mode']
-    assert isinstance(auto_mode, AutoMode)
-    await auto_mode.turn_on()
-    return True
-
-
-async def mutate_turn_off(info: Info) -> bool:
-    auto_mode = info.context['schedule']['auto_mode']
-    assert isinstance(auto_mode, AutoMode)
-    await auto_mode.turn_off()
-    return True
-
-
-@strawberry.type
-class MutationScheduleAutoMode:
-    turn_on: bool = strawberry.field(resolver=mutate_turn_on)
-    turn_off: bool = strawberry.field(resolver=mutate_turn_off)
 
 
 @strawberry.input
